add makefile
