+++ .gitignore
@@ -1,1 +1,5 @@
+# Ignore compiled files generated by GCC
 *.o
+*.out
+*.exe
+main.o
